Creacion

+++ src/main/java/cl/ecomarket/api/repository/ProductoRepository.java
@@ -3,4 +3,4 @@ package cl.ecomarket.api.repository;
 import cl.ecomarket.api.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Long> {}
+public interface ProductoRepository extends JpaRepository<Producto, Integer> {}
